organizing guideline lints to make them a bit more portable
