+++ clippy_lints/src/declared_lints.rs
@@ -158,7 +158,6 @@ pub(crate) static LINTS: &[&crate::LintInfo] = &[
     crate::exhaustive_items::EXHAUSTIVE_STRUCTS_INFO,
     crate::exit::EXIT_INFO,
     crate::explicit_write::EXPLICIT_WRITE_INFO,
-    crate::extern_without_repr::EXTERN_WITHOUT_REPR_INFO,
     crate::extra_unused_type_parameters::EXTRA_UNUSED_TYPE_PARAMETERS_INFO,
     crate::fallible_impl_from::FALLIBLE_IMPL_FROM_INFO,
     crate::float_literal::EXCESSIVE_PRECISION_INFO,
@@ -193,15 +192,18 @@ pub(crate) static LINTS: &[&crate::LintInfo] = &[
     crate::functions::TOO_MANY_LINES_INFO,
     crate::future_not_send::FUTURE_NOT_SEND_INFO,
     crate::guidelines::BLOCKING_OP_IN_ASYNC_INFO,
+    crate::guidelines::EXTERN_WITHOUT_REPR_INFO,
     crate::guidelines::FALLIBLE_MEMORY_ALLOCATION_INFO,
     crate::guidelines::MEM_UNSAFE_FUNCTIONS_INFO,
     crate::guidelines::PASSING_STRING_TO_C_FUNCTIONS_INFO,
     crate::guidelines::UNTRUSTED_LIB_LOADING_INFO,
     crate::guidelines::UNSAFE_BLOCK_IN_PROC_MACRO_INFO,
+    crate::guidelines_early::IMPLICIT_ABI_INFO,
+    crate::guidelines_early::LOOP_WITHOUT_BREAK_OR_RETURN_INFO,
+    crate::guidelines_early::NON_REENTRANT_FUNCTIONS_INFO,
     crate::if_let_mutex::IF_LET_MUTEX_INFO,
     crate::if_not_else::IF_NOT_ELSE_INFO,
     crate::if_then_some_else_none::IF_THEN_SOME_ELSE_NONE_INFO,
-    crate::implicit_abi::IMPLICIT_ABI_INFO,
     crate::implicit_hasher::IMPLICIT_HASHER_INFO,
     crate::implicit_return::IMPLICIT_RETURN_INFO,
     crate::implicit_saturating_add::IMPLICIT_SATURATING_ADD_INFO,
@@ -247,7 +249,6 @@ pub(crate) static LINTS: &[&crate::LintInfo] = &[
     crate::literal_representation::MISTYPED_LITERAL_SUFFIXES_INFO,
     crate::literal_representation::UNREADABLE_LITERAL_INFO,
     crate::literal_representation::UNUSUAL_BYTE_GROUPINGS_INFO,
-    crate::loop_without_break_or_return::LOOP_WITHOUT_BREAK_OR_RETURN_INFO,
     crate::loops::EMPTY_LOOP_INFO,
     crate::loops::EXPLICIT_COUNTER_LOOP_INFO,
     crate::loops::EXPLICIT_INTO_ITER_LOOP_INFO,
@@ -475,7 +476,6 @@ pub(crate) static LINTS: &[&crate::LintInfo] = &[
     crate::non_expressive_names::MANY_SINGLE_CHAR_NAMES_INFO,
     crate::non_expressive_names::SIMILAR_NAMES_INFO,
     crate::non_octal_unix_permissions::NON_OCTAL_UNIX_PERMISSIONS_INFO,
-    crate::non_reentrant_functions::NON_REENTRANT_FUNCTIONS_INFO,
     crate::non_send_fields_in_send_ty::NON_SEND_FIELDS_IN_SEND_TY_INFO,
     crate::nonstandard_macro_braces::NONSTANDARD_MACRO_BRACES_INFO,
     crate::octal_escapes::OCTAL_ESCAPES_INFO,
