+++ tests/ui/guidelines/blocking_op_in_async.stderr
