+++ tests/ui/guidelines/extern_without_repr.rs
