+++ tests/ui/guidelines/falliable_memory_allocation.rs
