+++ tests/ui/guidelines/loop_without_break_or_return.rs
