+++ tests/ui/guidelines/non_reentrant_functions.stderr
