+++ tests/ui/guidelines/passing_string_to_c_functions.stderr
