+++ tests/ui/guidelines/unsafe_block_in_proc_macro.rs
