+++ tests/ui/guidelines/untrusted_lib_loading.rs
